// src/shared/types/movie.ts
export interface MovieSearchResult {
  Title: string
  Year: string
  imdbID: string
  Type: 'movie' | 'series' | 'episode'
  Poster: string
}

export interface MovieSearchResponse {
  Search: MovieSearchResult[]
  totalResults: string
  Response: 'True' | 'False'
  Error?: string
}

export interface MovieDetail {
  Title: string
  Year: string
  Rated: string
  Released: string
  Runtime: string
  Genre: string
  Director: string
  Writer: string
  Actors: string
  Plot: string
  Language: string
  Country: string
  Awards: string
  Poster: string
  Ratings: { Source: string; Value: string }[]
  Metascore: string
  imdbRating: string
  imdbVotes: string
  imdbID: string
  Type: string
  DVD: string
  BoxOffice: string
  Production: string
  Website: string
  Response: 'True' | 'False'
  Error?: string
}

export interface SearchParams {
  query: string
  page: number
  type?: 'movie' | 'series' | ''
  year?: string
}